+++ src/utils/service/service.js
@@ -1,5 +1,6 @@
 import axios from 'axios';
 import { assign } from 'lodash';
+import Logger from '../logger';
 
 /**
  * Global configuration for all service classes.
@@ -164,15 +165,31 @@ class Service {
    *
    * @method get
    * @param {Number} id - the ID of the resource.
-   * @param {Function} onSuccess - a callback to trigger on success
-   * @param {Function} onError - a callback to trigger on error
+   * @param {Object} options - an object with options for query params and success/error callbacks
+   * @param {Function} _onError - [Deprecated] a callback to trigger on error
    * @return {Void}
    */
-  get(id, onSuccess, onError) {
-    this.client.get(String(id)).then(
-      this.handleResponse.bind(this, onSuccess),
-      this.handleResponse.bind(this, onError)
-    );
+  get(id, options, _onError) {
+    let request;
+
+    if (typeof (options) === 'function') {
+      Logger.deprecate('Passing onSuccess and onError as separate arguments is deprecated. Please use `get(1, { onSuccess: function, onError: function, params: { k: "v" } })`'); // eslint-disable-line max-len
+
+      request = this.client.get(String(id));
+
+      request.then(
+        this.handleResponse.bind(this, options),
+        this.handleResponse.bind(this, _onError)
+      );
+    } else {
+      const { onSuccess, onError, ...params } = options;
+
+      request = this.client.get(String(id), params);
+      request.then(
+        this.handleResponse.bind(this, onSuccess),
+        this.handleResponse.bind(this, onError)
+      );
+    }
   }
 
   /**
@@ -180,15 +197,31 @@ class Service {
    *
    * @method post
    * @param {Object} data - the data to post to the server.
-   * @param {Function} onSuccess - a callback to trigger on success
-   * @param {Function} onError - a callback to trigger on error
+   * @param {Object} options - an object with options for query params and success/error callbacks
+   * @param {Function} _onError - [Deprecated] a callback to trigger on error
    * @return {Void}
    */
-  post(data, onSuccess, onError) {
-    this.client.post('', data).then(
-      this.handleResponse.bind(this, onSuccess),
-      this.handleResponse.bind(this, onError)
-    );
+  post(data, options, _onError) {
+    let request;
+
+    if (typeof (options) === 'function') {
+      Logger.deprecate("Passing onSuccess and onError as separate arguments is deprecated. Please use `post('', { data_key: 'data_val' }, { onSuccess: function, onError: function, params: { k: 'v' } })`"); // eslint-disable-line max-len
+      request = this.client.post('', data);
+
+      request.then(
+        this.handleResponse.bind(this, options),
+        this.handleResponse.bind(this, _onError)
+      );
+    } else {
+      const { onSuccess, onError, ...params } = options;
+
+      request = this.client.post('', data, params);
+
+      request.then(
+        this.handleResponse.bind(this, onSuccess),
+        this.handleResponse.bind(this, onError)
+      );
+    }
   }
 
   /**
@@ -197,15 +230,31 @@ class Service {
    * @method put
    * @param {Number} id - the ID of the resource.
    * @param {Object} data - the data to post to the server.
-   * @param {Function} onSuccess - a callback to trigger on success
-   * @param {Function} onError - a callback to trigger on error
+   * @param {Object} options - an object with options for query params and success/error callbacks
+   * @param {Function} _onError - [Deprecated] a callback to trigger on error
    * @return {Void}
    */
-  put(id, data, onSuccess, onError) {
-    this.client.put(String(id), data).then(
-      this.handleResponse.bind(this, onSuccess),
-      this.handleResponse.bind(this, onError)
-    );
+  put(id, data, options, _onError) {
+    let request;
+
+    if (typeof (options) === 'function') {
+      Logger.deprecate("Passing onSuccess and onError as separate arguments is deprecated. Please use `put(1, { data_key: 'data_val' }, { onSuccess: function, onError: function, params: { k: 'v' } })`"); // eslint-disable-line max-len
+      request = this.client.put(String(id), data);
+
+      request.then(
+        this.handleResponse.bind(this, options),
+        this.handleResponse.bind(this, _onError)
+      );
+    } else {
+      const { onSuccess, onError, ...params } = options;
+
+      request = this.client.put(String(id), data, params);
+
+      request.then(
+        this.handleResponse.bind(this, onSuccess),
+        this.handleResponse.bind(this, onError)
+      );
+    }
   }
 
   /**
@@ -213,15 +262,31 @@ class Service {
    *
    * @method delete
    * @param {Number} id - the ID of the resource.
-   * @param {Function} onSuccess - a callback to trigger on success
-   * @param {Function} onError - a callback to trigger on error
+   * @param {Object} options - an object with options for query params and success/error callbacks
+   * @param {Function} _onError - [Deprecated] a callback to trigger on error
    * @return {Void}
    */
-  delete(id, onSuccess, onError) {
-    this.client.delete(String(id)).then(
-      this.handleResponse.bind(this, onSuccess),
-      this.handleResponse.bind(this, onError)
-    );
+  delete(id, options, _onError) {
+    let request;
+
+    if (typeof (options) === 'function') {
+      Logger.deprecate("Passing onSuccess and onError as separate arguments is deprecated. Please use `delete(1, { onSuccess: function, onError: function, params: { k: 'v' } })`"); // eslint-disable-line max-len
+      request = this.client.delete(String(id));
+
+      request.then(
+        this.handleResponse.bind(this, options),
+        this.handleResponse.bind(this, _onError)
+      );
+    } else {
+      const { onSuccess, onError, ...params } = options;
+
+      request = this.client.delete(String(id), params);
+
+      request.then(
+        this.handleResponse.bind(this, onSuccess),
+        this.handleResponse.bind(this, onError)
+      );
+    }
   }
 
   /**